+++ lib/utils/pubspec_updater.dart
@@ -4,6 +4,7 @@ const Map<String, String> packageVersions = {
   'flutter_bloc': '^8.1.0',
   'equatable': '^2.0.5',
   'flutter_riverpod': '^2.4.0',
+  'get': '^4.6.6',
   'dartz': '^0.10.1',
   'get_it': '^7.6.0',
   'http': '^1.1.0',